feat: Ability to add custom NPM repository scopes.
While the prompt did exist to add a custom NPM repository scope there
was nowhere for a client to enter such information nor was any of the
information being used if it was submitted (Essentially it was useless).

The prompt has been updated to encourage the client to enter their
custom scope and it will actually be reflected in the `package.json`.

Signed-off-by: staticBanter <76642517+staticBanter@users.noreply.github.com>